// wardn/frontend/lib/api/generated/model/mCPGatewayToolApprovalRead.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.0.1
 */
import type { MCPGatewayToolApprovalReadArguments } from './mCPGatewayToolApprovalReadArguments';
import type { MCPGatewayToolApprovalReadGuardrail } from './mCPGatewayToolApprovalReadGuardrail';
import type { MCPGatewayToolApprovalReadRequestMeta } from './mCPGatewayToolApprovalReadRequestMeta';
import type { MCPGatewayToolApprovalReadResult } from './mCPGatewayToolApprovalReadResult';

export interface MCPGatewayToolApprovalRead {
  arguments: MCPGatewayToolApprovalReadArguments;
  createdAt: string;
  decidedById?: string | null;
  error?: string;
  guardrail: MCPGatewayToolApprovalReadGuardrail;
  id: string;
  installationId: string;
  organizationId: string;
  requestMeta: MCPGatewayToolApprovalReadRequestMeta;
  requestedById?: string | null;
  result?: MCPGatewayToolApprovalReadResult;
  serverName: string;
  status: string;
  toolCallId: string;
  toolName: string;
  toolSchemaId?: string | null;
  updatedAt: string;
  workspaceId: string;
}